fix(core): use usize for subscription id in text binder self-unsubscribe
The PR #218 sweep that switched Signal::subscribe and Signal::unsubscribe
to usize missed this third call site in core/src/vdom/cast/impl.rs
where the binder closure for an as_reactive_text node stores the
subscription id in an Rc<Cell<u64>> so the closure can call
source.unsubscribe(listener_id.get()) from inside itself when the
target text node detaches.

After the Signal API change the listener id is a usize but the binder
local was still typed as u64 and read back as u64, so
source.unsubscribe(listener_id.get()) stopped compiling with:

  error[E0308]: mismatched types
    --> core/src/vdom/cast/impl.rs:259:40
      |
    259 |                     source.unsubscribe(listener_id.get());
      |                            ----------- ^^^^^^^^^^^^^^^^^ expected `usize`, found `u64`

This slipped past the PR #218 local cargo check because the dev build
profile compiles a different subset and the release+wasm32 target was
not exercised locally. The CI run on master after the merge caught it.

Switch the three Cell types and the initial sentinel to usize::MAX so
the binder matches the listener id the rest of the reactive subsystem
hands out.

+++ core/src/vdom/cast/impl.rs
@@ -252,9 +252,9 @@ where
         let source: Signal<T> = *self;
         let binder: Rc<dyn Fn(&Text)> = Rc::new(move |text: &Text| {
             let text_node: Text = text.clone();
-            let subscription_id: Rc<Cell<u64>> = Rc::new(Cell::new(u64::MAX));
-            let listener_id: Rc<Cell<u64>> = subscription_id.clone();
-            let id: u64 = source.subscribe(move || {
+            let subscription_id: Rc<Cell<usize>> = Rc::new(Cell::new(usize::MAX));
+            let listener_id: Rc<Cell<usize>> = subscription_id.clone();
+            let id: usize = source.subscribe(move || {
                 if !text_node.is_connected() {
                     source.unsubscribe(listener_id.get());
                     return;
